Refactor to combine Dual & single category files
