+++ src/api/requests/downloads/getDownloads.spec.ts
@@ -3,7 +3,21 @@ import { logger } from '@/lib/logger'
 import { downloadsCsvFixture } from '@/mock-server/handlers/downloads/fixtures/downloads-csv'
 import { downloadsJsonFixture } from '@/mock-server/handlers/downloads/fixtures/downloads-json'
 
-import { getDownloads } from './getDownloads'
+import { getDownloads, SingleCategoryRequestParams } from './getDownloads'
+
+const basePlot: SingleCategoryRequestParams['plots'][number] = {
+  topic: 'COVID-19',
+  metric: 'new_cases_7days_sum',
+  stratum: '',
+}
+
+const baseBody: SingleCategoryRequestParams = {
+  is_public: true,
+  file_format: 'csv',
+  x_axis: null,
+  confidence_intervals: false,
+  plots: [basePlot],
+}
 
 test('Returns chart data in CSV format', async () => {
   jest.mocked(client).mockResolvedValueOnce({
@@ -11,13 +25,7 @@ test('Returns chart data in CSV format', async () => {
     status: 200,
   })
 
-  const result = await getDownloads(true, [
-    {
-      topic: 'COVID-19',
-      metric: 'new_cases_7days_sum',
-      stratum: '',
-    },
-  ])
+  const result = await getDownloads(baseBody)
 
   expect(result).toEqual(downloadsCsvFixture)
 })
@@ -28,17 +36,7 @@ test('Returns chart data in json format', async () => {
     status: 200,
   })
 
-  const result = await getDownloads(
-    true,
-    [
-      {
-        topic: 'COVID-19',
-        metric: 'new_cases_7days_sum',
-        stratum: '',
-      },
-    ],
-    'json'
-  )
+  const result = await getDownloads({ ...baseBody, file_format: 'json' })
 
   expect(result).toEqual(downloadsJsonFixture)
 })
@@ -49,17 +47,7 @@ test('Handles generic http errors', async () => {
     status: 400,
   })
 
-  const result = await getDownloads(
-    true,
-    [
-      {
-        topic: 'COVID-19',
-        metric: 'new_cases_7days_sum',
-        stratum: '',
-      },
-    ],
-    'json'
-  )
+  const result = await getDownloads({ ...baseBody, file_format: 'json' })
 
   expect(logger.error).toHaveBeenCalledTimes(1)
 
@@ -72,30 +60,12 @@ test('Sends confidence_intervals as true in request body', async () => {
     status: 200,
   })
 
-  await getDownloads(
-    true,
-    [
-      {
-        topic: 'COVID-19',
-        metric: 'new_cases_7days_sum',
-        stratum: '',
-      },
-    ],
-    'csv',
-    null,
-    true
-  )
+  await getDownloads({ ...baseBody, confidence_intervals: true })
 
   expect(client).toHaveBeenCalledWith('downloads/v2', {
     body: {
       is_public: true,
-      plots: [
-        {
-          topic: 'COVID-19',
-          metric: 'new_cases_7days_sum',
-          stratum: '',
-        },
-      ],
+      plots: [basePlot],
       file_format: 'csv',
       x_axis: null,
       confidence_intervals: true,
@@ -109,30 +79,12 @@ test('Sends confidence_intervals as false in request body', async () => {
     status: 200,
   })
 
-  await getDownloads(
-    true,
-    [
-      {
-        topic: 'COVID-19',
-        metric: 'new_cases_7days_sum',
-        stratum: '',
-      },
-    ],
-    'csv',
-    null,
-    false
-  )
+  await getDownloads({ ...baseBody, confidence_intervals: false })
 
   expect(client).toHaveBeenCalledWith('downloads/v2', {
     body: {
       is_public: true,
-      plots: [
-        {
-          topic: 'COVID-19',
-          metric: 'new_cases_7days_sum',
-          stratum: '',
-        },
-      ],
+      plots: [basePlot],
       file_format: 'csv',
       x_axis: null,
       confidence_intervals: false,
@@ -146,24 +98,12 @@ test('Defaults confidence_intervals to false when not provided', async () => {
     status: 200,
   })
 
-  await getDownloads(true, [
-    {
-      topic: 'COVID-19',
-      metric: 'new_cases_7days_sum',
-      stratum: '',
-    },
-  ])
+  await getDownloads(baseBody)
 
   expect(client).toHaveBeenCalledWith('downloads/v2', {
     body: {
       is_public: true,
-      plots: [
-        {
-          topic: 'COVID-19',
-          metric: 'new_cases_7days_sum',
-          stratum: '',
-        },
-      ],
+      plots: [basePlot],
       file_format: 'csv',
       x_axis: null,
       confidence_intervals: false,
