+++ src/api/requests/downloads/getDownloads.ts
@@ -1,11 +1,24 @@
 import { z } from 'zod'
 
 import { Geography, GeographyType, Metrics, Topics } from '@/api/models'
+import { ChartLineColours, ChartTypes } from '@/api/models/Chart'
 import { client } from '@/api/utils/api.utils'
 import { auth } from '@/auth'
 import { auditLog, logger } from '@/lib/logger'
 
-export const requestSchema = z.object({
+const dualCategoryStaticFieldsSchema = z.object({
+  topic: Topics,
+  metric: Metrics,
+  geography: Geography.optional(),
+  geography_type: GeographyType.optional(),
+  sex: z.string().nullable().optional(),
+  age: z.string().nullable().optional(),
+  stratum: z.string().optional(),
+  date_from: z.string().nullable().optional(),
+  date_to: z.string().nullable().optional(),
+})
+
+export const singleCategoryRequestSchema = z.object({
   is_public: z.boolean().default(true),
   file_format: z.enum(['json', 'csv']),
   x_axis: z.enum(['age', 'geography', 'sex', 'stratum', 'date', 'metric']).optional().nullable(),
@@ -24,32 +37,47 @@ export const requestSchema = z.object({
     })
   ),
 })
+export const requestSchema = singleCategoryRequestSchema
 
-export type RequestParams = z.infer<typeof requestSchema>
+export const dualCategoryRequestSchema = z.object({
+  is_public: z.boolean().default(true),
+  file_format: z.enum(['json', 'csv']),
+  x_axis: z.string().nullable().optional(),
+  y_axis: z.string().nullable().optional(),
+  x_axis_title: z.string().optional(),
+  y_axis_title: z.string().optional(),
+  y_axis_minimum_value: z.number().nullable().optional(),
+  y_axis_maximum_value: z.number().nullable().optional(),
+  chart_type: ChartTypes,
+  static_fields: dualCategoryStaticFieldsSchema,
+  primary_field_values: z.array(z.string()),
+  secondary_category: z.string(),
+  segments: z.array(
+    z.object({
+      secondary_field_value: z.string(),
+      colour: ChartLineColours,
+      label: z.string().nullable().optional(),
+    })
+  ),
+})
 
-export const getDownloads = async (
-  is_public: RequestParams['is_public'],
-  plots: RequestParams['plots'],
-  format: RequestParams['file_format'] = 'csv',
-  x_axis: RequestParams['x_axis'] = null,
-  confidence_intervals: RequestParams['confidence_intervals'] = false
-) => {
+export type SingleCategoryRequestParams = z.infer<typeof singleCategoryRequestSchema>
+export type DualCategoryRequestParams = z.infer<typeof dualCategoryRequestSchema>
+export type RequestParams = SingleCategoryRequestParams | DualCategoryRequestParams
+
+export const getDownloads = async (body: RequestParams) => {
   try {
-    if (!is_public) {
+    if (!body.is_public) {
       const session = await auth()
       if (session) {
-        auditLog(session.userId ?? '', 'FILE_DOWNLOAD', `${format} - ${JSON.stringify(plots)}`)
+        const auditDetail = 'plots' in body ? JSON.stringify(body.plots) : JSON.stringify(body.static_fields)
+        auditLog(session.userId ?? '', 'FILE_DOWNLOAD', `${body.file_format} - ${auditDetail}`)
       }
     }
 
-    const body: RequestParams = {
-      is_public,
-      plots,
-      x_axis,
-      file_format: format,
-      confidence_intervals,
-    }
-    const { data } = await client<string>(`downloads/v2`, { body })
+    const isDualCategory = 'static_fields' in body
+    const endpoint = isDualCategory ? `downloads/dual-category/v1` : `downloads/v2`
+    const { data } = await client<string>(endpoint, { body })
 
     return data
   } catch (error) {
